[messages] moved messages to pagination component
